run all tests as part of build-app.sh
