+++ app.Dockerfile
@@ -13,6 +13,6 @@ ENV APP_DOCKER_BRANCH_TAG $registry/$name:$git_branch
 RUN mkdir -p /etc/orderly/web
 RUN touch /etc/orderly/web/go_signal
 
-CMD ./gradlew :app:distDocker -i -Pdocker_version=$GIT_ID -Pdocker_name=$APP_DOCKER_TAG \
+CMD ./gradlew :app:test :customConfigTests:test :app:distDocker -i -Pdocker_version=$GIT_ID -Pdocker_name=$APP_DOCKER_TAG \
     && docker tag $APP_DOCKER_COMMIT_TAG $APP_DOCKER_BRANCH_TAG \
     && docker push $APP_DOCKER_BRANCH_TAG
